Fix log type casing inconsistency and ChatDogProfile.age type
- Fix ChatDogProfile.age from Int? to Double? to match Dog.age type change
- Normalize log types in handleLogSuggestion to match LogType enum (capitalized)
- Fix DailyHealthReviewView to use capitalized log types (Meals, Walk, Playtime, Water)
- Fix SmartInsightsView to use capitalized log types throughout

This ensures AI-logged meals/water appear correctly in Daily Health Review and Health Insights.

+++ ios-app/DogHealthApp/DogHealthApp/Services/APIService.swift
@@ -343,7 +343,7 @@ struct ChatRequest: Codable {
 struct ChatDogProfile: Codable {
     let name: String
     let breed: String?
-    let age: Int?
+    let age: Double?
     let weight: Double?
     let healthConcerns: [String]?
     let allergies: [String]?
